+++ .config/javascript/config.js
@@ -1,0 +1,7 @@
+export const npmScripts = [
+  { name: 'build:gas', cmd: 'vite build --config vite.config.gas.js' },
+  { name: 'build:cp', cmd: 'vite build --config vite.config.copy-paste.js' },
+  { name: 'build:node', cmd: 'vite build --config vite.config.node.js' },
+];
+
+export const deps = null;