+++ src/client/lib/mocking/index.js
@@ -1,0 +1,3 @@
+import polyfillScriptRun from './polyfill-script-run.js';
+
+export { polyfillScriptRun };