+++ src/client/mocks/export-test.js
@@ -1,7 +1,0 @@
-export function test1() {
-  console.log('1');
-}
-
-export function test2() {
-  console.log('2');
-}